Add dependency management to UILayoutProxy
This commit introduces dependency management in the UILayoutProxy class. A new feature was implemented where a ViewModel can have dependencies which are automatically hidden when the ViewModel itself is hidden. New methods have also been added to allow showing a ViewModel as a dependency of another. Additionally, the package version has been updated to 1.0.16.